pyproject.toml
[project]
name = "beyond-chatgpt"
version = "0.2.0"
description = "gpt-like-a-dev (lightweight version optimized for Vercel deployment)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core API dependencies
    "fastapi>=0.116.2",
    "uvicorn>=0.35.0",
    "openai>=1.107.3",
    "together>=0.2.7",
    "ollama>=0.4.4",
    "pydantic>=2.11.9",
    "pydantic-core>=2.33.2",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "starlette>=0.48.0",
    # Lightweight PDF processing dependencies
    "PyPDF2>=3.0.1",
    "pdfplumber>=0.9.0",
    "tiktoken>=0.11.0",
    "python-docx>=1.1.2",
    "python-pptx>=0.6.23",
    # Supporting dependencies
    "httpx>=0.28.1",
    "httpcore>=1.0.9",
    "requests>=2.32.5",
    "python-dotenv>=1.1.1",
    # Jupyter and development dependencies
    "jupyter>=1.1.1",
    "ipykernel>=6.29.5",
    "numpy>=2.3.3",
]
